hog: HID I/O driver
uHID is HID I/O driver that makes possible to implement HID I/O drivers
in user-space. It works similar to the uinput but it is initialized with
a HID descriptor and deals with raw HID reports.

This commit uses uHID to create a HID device for the remote HoG device
and to tranfers HID reports to HID subsystem.