hcireplay: Rework packet handling and improve output
Add handling of type configurations. process_in() now applies type
config to all incoming packets (including unknowns), which allows
to modify handling of packets that are not in the sequence.

Add get_type_attr() to return the type configuration for a given
frame. It also returns the type config of the according opcode
for "Command Complete" events so that they are no longer replayed
when their command packets are skipped or emulated.

Check for match in process_in() and handle unknowns separately.

Modify process() to only move to the next packet in the sequence
when it has been processed. Handling unknowns in process_in() will
not move the sequence forward.

Replace all replay_ functions with direct calls to send_frm() and
dump_frame().

Improve output with minor changes and by introducing consistent tags
for packets in the sequence, packets handled by the emulator and
unknowns.

Move the call of vhci_open() so the interface is initialised as
late as possible. This should improve timing and prevent loss of
packets.